Fix multiblocks using day time to advance their renders, not game time

+++ Common/src/main/java/vazkii/patchouli/api/IStateMatcher.java
@@ -14,10 +14,10 @@ public interface IStateMatcher {
 	/**
 	 * Gets the state displayed by this state matcher for rendering
 	 * the multiblock page type and the in-world preview.
-	 * 
+	 *
 	 * @param ticks World ticks, to allow cycling the state shown.
 	 */
-	BlockState getDisplayedState(int ticks);
+	BlockState getDisplayedState(long ticks);
 
 	/**
 	 * Returns a predicate that validates whether the given state is
