+++ Common/src/main/java/vazkii/patchouli/api/stub/StubMatcher.java
@@ -17,7 +17,7 @@ public final class StubMatcher implements IStateMatcher {
 	private StubMatcher() {}
 
 	@Override
-	public BlockState getDisplayedState(int ticks) {
+	public BlockState getDisplayedState(long ticks) {
 		return state;
 	}
